+++ CHANGES.md
@@ -1,3 +1,3 @@
-* Layout to "circo"
-* Head label font size down to 8
-* 
+* Added basic_import_yaml.py: Imports a YAML file with the subnet and device info.
+* Updated README.md with more options
+* Added basic_import_netbox.py: Imports subnet and device info from Netbox. INCOMPLETE!!!!